// frontend/src/sections/plugin-management/plugins.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../api/graphql";

export type InstalledPluginsVariables = Record<string, never>;

export type InstalledPluginsResult = {
  centralServer: {
  plugin: {
  installedPlugins: {
  totalCount: number;
  nodes: Array<{
  id: string;
  code: string;
  version: string;
  kind: "BACKEND" | "FRONTEND";
  types: Array<string>;
  hostRuntime: string | null;
}>;
};
};
};
};

export const InstalledPlugins = {
  query: "query InstalledPlugins {\n  centralServer {\n    plugin {\n      installedPlugins {\n        totalCount\n        nodes {\n          id\n          code\n          version\n          kind\n          types\n          hostRuntime\n        }\n      }\n    }\n  }\n}",
} as TypedDocument<InstalledPluginsResult, InstalledPluginsVariables>;

export type InstallUploadedPluginVariables = {
  fileId: string;
};

export type InstallUploadedPluginResult = {
  centralServer: {
  plugins: {
  installUploadedPlugin: {
  pluginInfo: unknown;
};
};
};
};

export const InstallUploadedPlugin = {
  query: "mutation InstallUploadedPlugin($fileId: String!) {\n  centralServer {\n    plugins {\n      installUploadedPlugin(fileId: $fileId) {\n        pluginInfo\n      }\n    }\n  }\n}",
} as TypedDocument<InstallUploadedPluginResult, InstallUploadedPluginVariables>;

export type UninstallPluginVariables = {
  id: string;
};

export type UninstallPluginResult = {
  centralServer: {
  plugins: {
  uninstallPlugin: {
  id: string;
  code: string;
  kind: "BACKEND" | "FRONTEND";
};
};
};
};

export const UninstallPlugin = {
  query: "mutation UninstallPlugin($id: String!) {\n  centralServer {\n    plugins {\n      uninstallPlugin(id: $id) {\n        id\n        code\n        kind\n      }\n    }\n  }\n}",
} as TypedDocument<UninstallPluginResult, UninstallPluginVariables>;

export type PluginConfigurationVariables = {
  storeId: string;
  pluginCode: string;
};

export type PluginConfigurationResult = {
  pluginData: ({
  nodes: Array<{
  id: string;
  storeId: string | null;
  data: string;
}>;
});
};

export const PluginConfiguration = {
  query: "query PluginConfiguration($storeId: String!, $pluginCode: String!) {\n  pluginData(\n    storeId: $storeId\n    pluginCode: $pluginCode\n    filter: {dataIdentifier: {equalTo: \"configuration\"}}\n  ) {\n    ... on PluginDataConnector {\n      nodes {\n        id\n        storeId\n        data\n      }\n    }\n  }\n}",
} as TypedDocument<PluginConfigurationResult, PluginConfigurationVariables>;

export type InsertPluginConfigurationVariables = {
  storeId: string;
  input: {
    id: string;
    storeId?: string | null;
    pluginCode: string;
    relatedRecordId?: string | null;
    dataIdentifier: string;
    data: string;
    datetime?: string | null;
  };
};

export type InsertPluginConfigurationResult = {
  insertPluginData: ({
  id: string;
});
};

export const InsertPluginConfiguration = {
  query: "mutation InsertPluginConfiguration($storeId: String!, $input: InsertPluginDataInput!) {\n  insertPluginData(storeId: $storeId, input: $input) {\n    ... on PluginDataNode {\n      id\n    }\n  }\n}",
} as TypedDocument<InsertPluginConfigurationResult, InsertPluginConfigurationVariables>;

export type UpdatePluginConfigurationVariables = {
  storeId: string;
  input: {
    id: string;
    storeId?: string | null;
    pluginCode: string;
    relatedRecordId?: string | null;
    dataIdentifier: string;
    data: string;
    datetime?: string | null;
  };
};

export type UpdatePluginConfigurationResult = {
  updatePluginData: ({
  id: string;
});
};

export const UpdatePluginConfiguration = {
  query: "mutation UpdatePluginConfiguration($storeId: String!, $input: UpdatePluginDataInput!) {\n  updatePluginData(storeId: $storeId, input: $input) {\n    ... on PluginDataNode {\n      id\n    }\n  }\n}",
} as TypedDocument<UpdatePluginConfigurationResult, UpdatePluginConfigurationVariables>;